Always defer to private/public keys on filesystem
If using a persistent volume that will end up migrating from node to
node, such as an EBS volume on EC2, then the cookbook will now always
defer to the files on disk and not overwrite them. It will also update
the node pubkey setting from the data on disk.